+++ Models/OdometerRecord/OdometerRecordInput.cs
@@ -5,8 +5,8 @@
         public int Id { get; set; }
         public int VehicleId { get; set; }
         public string Date { get; set; } = DateTime.Now.ToShortDateString();
-        public int InitialMileage { get; set; }
-        public int Mileage { get; set; }
+        public decimal InitialMileage { get; set; }
+        public decimal Mileage { get; set; }
         public string Notes { get; set; }
         public List<UploadedFiles> Files { get; set; } = new List<UploadedFiles>();
         public List<string> Tags { get; set; } = new List<string>();
